Add support for lazy #map and #select on cursors
Db#each without a block returns an enumerable that will perform the same iteration. You can call #map, #select and other Enumerable methods on this object and the block given to these methods will operate on each key/value pair as it is loaded. The result will be an Array. If you chain a number of operations all but the first will operate on a materialized array, which might not be very efficient.

In Ruby 2.0 this is solved by Enumerable::Lazy, but in JRuby 1.7.x that class is not available.

This code makes #map and #select return enumerables that will evaluate those two methods lazily until any other Enumerable method is used.