Rename Cursor -> Iterator
I was reluctant to use the name Iterator since it clashes with java.util.Iterator, but I've decided now that I don't think it will become a problem. I want to write documentation and I would prefer it to use the same nomenclature as the RocksDB docs, so having the right name for things seems like a good idea.